+++ libft/libft.h
@@ -47,7 +47,8 @@ int		ft_tolower(int c);
 
 int		ft_strncmp(const char *str1, const char *str2, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
-int		ft_atoi(const char *str);
+
+long	ft_atoi(const char *str);
 
 size_t	ft_strlen(const char *str);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
